new scripts for firebase auth
created functions for signin, signout, added observer for auth state, delete user, update email, password reset and reauthenticate. A separate handler for signing up or in needs to be added ( could be a button or a whole new form ), pages or components for updating users email, a button for deleting the current logged in user. also someway to enter an email to send a password reset link to